third_party/WebKit/LayoutTests/platform/linux/http/tests/w3c/webperf/submission/Intel/resource-timing/test_resource_timing_buffer_size_restriction-expected.txt

This is a testharness.js-based test.
FAIL 2 resource timing entries should be stored since buffer size is 2! assert_equals: 2 resource timing entries should be stored since buffer size is 2! expected 2 but got 4
Harness: the test ran to completion.
